Fix bug in searchablehelp topics
Search result would not always expand due to an issue of an Accordion being instantiated more than once for the search accordion, causing each click to fire twice (creating a rapid open/close). This solution removes it from the DOM at the time that the USWDS code looks for acordions to instantiate, so it only gets instantiated once (by us).