Implement desired Summary screen functionality
Still needs UI tweaking, but the logic and components should be there.

Note that there's a use case unaccounted for by the specs in #40: when there are no household programs and household income is not collected (which happens if all students are foster, or if all are FHMR but income reporting is not elected). This needs to be addressed.