Add App-level IntlProvider
This sets us up to be able to remove IntlProviders from around each FormattedMessage (and thereby deprecate our own FormattedMessage components) so that babel-plugin-react-intl begins to function properly again.

One implication of this is that hot-swapping locales will now cause the entire DOM to refresh (rather than each individual message), which requires us to re-navigate to the current slide after the DOM re-renders.